ch(Authentication): Add authentication to route [Start #167239453]
