+++ .env
@@ -4,4 +4,4 @@ PGDATABASE=wayfarerapi
 PGPASSWORD=bolatito92
 PGPORT=5432
 
-SECRET=AKDFJAKJDFKSJADKAJKASJKLLW13408954
+SECRET=olaoyeadewale
