+++ src/controller/bus.controller.js
@@ -1,9 +1,98 @@
-const createBusQuery = `INSERT INTO 
-bus (number_plate, manufacturer, model, year, capacity, created_on) 
-      VALUES($1, $2, $3, $4, $5, $6) 
-      returning *`;
+/* eslint-disable camelcase */
+import db from '../model/db';
+import { createBusQuery, getAllBusQuery } from '../model/bus.model';
 
-const busAvailability = 'SELECT * FROM trip WHERE (trip_date = $1 AND bus_id = $2 AND status = $3)';
+const Bus = {
+  /**
+* Admin can add bus
+* @param {object} req
+* @param {object} res
+* @returns {object} bus object
+*/
 
-// eslint-disable-next-line import/prefer-default-export
-export { createBusQuery, busAvailability };
+  async addBus(req, res) {
+    const {
+      number_plate, manufacturer, model, year, capacity,
+    } = req.body;
+
+
+    if (!number_plate || !capacity) {
+      return res.status(400).json({
+        status: 'error',
+        error: 'Plate Number and Capacity is reaquired',
+      });
+    }
+    const values = [
+      number_plate,
+      manufacturer,
+      model,
+      year,
+      capacity,
+      new Date(),
+    ];
+
+    try {
+      if (req.admin === false) {
+        return res.status(403).json({
+          status: 'error',
+          error: 'Unauthorized!',
+        });
+      }
+      const { rows } = await db.query(createBusQuery, values);
+      const { bus_id } = rows[0];
+
+      return res.status(201).json({
+        status: 'suceess',
+        data: {
+          bus_id,
+          number_plate,
+          manufacturer,
+          model,
+          year,
+          capacity,
+        },
+      });
+    } catch (error) {
+      if (error.routine === '_bt_check_unique') {
+        return res.status(409).json({
+          status: 'error',
+          error: 'Bus already exist',
+        });
+      }
+      return res.status(400).json({
+        status: 'error',
+        error: 'Something went wrong, contact our engineers',
+      });
+    }
+  },
+
+  /**
+    * Admin can see all buses
+    * @param {object} req
+    * @param {object} res
+    * @returns {object} bus object
+    */
+  // eslint-disable-next-line consistent-return
+  async getAllBus(req, res) {
+    try {
+      if (req.admin === false) {
+        return res.status(403).json({
+          status: 'error',
+          error: 'Unauthorized!',
+        });
+      }
+      const { rows } = await db.query(getAllBusQuery);
+      return res.status(200).json({
+        status: 'success',
+        data: rows,
+      });
+    } catch (error) {
+      return res.status(400).json({
+        status: 'error',
+        error: 'Something went wrong, contact our engineers',
+      });
+    }
+  },
+};
+
+export default Bus;
